fix(playback): split tests.rs to stay under 1000-line cap
The PlaybackEngineConfig refactor (#59 in this PR) added ~25 lines
across tests.rs by widening every test's PlaybackEngine::new(...) call
into a struct-literal block, pushing the file from 996 to 1025 lines —
over the airuleset cap, caught by the File Size Check job.

Extracted the handle_previous + processed-event tests to a new sibling
tests_history.rs (#[path]-included from mod.rs). Same pattern as the
other tests_* siblings in this directory (tests_play_video,
tests_scene_change, tests_song_end).

tests.rs now 609 lines, tests_history.rs is 424.
